backend/pyproject.toml
[tool.poetry]
name = "chessica-backend"
version = "0.1.0"
description = "Backend API for the Chessica exploit-aware chess engine"
authors = ["Chessica Team"]
readme = "../README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
redis = "^5.0.4"
python-chess = "^1.999"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
